// packages/astrocite-eutils/src/schema.d.ts
/**
 * Format: "<YYYY> <Mmm>? <DD>?"
 * Examples: `2004`, `2014 Jun`, `2000 Nov 12`
 */
type ShortDate = string;

/**
 * Format: "YYYY/MM/DD HH:MM"
 * Example: `2014/07/01 00:00`
 */
type LongDate = string;

interface Reference {
    refsource: string;
    reftype: string;
    pmid: number;
    note: string;
}

interface HistoryEntry {
    pubstatus: string;
    date: LongDate;
}

interface ArticleID {
    idtype: 'doi' | 'eid' | 'pii' | 'pmc' | 'pubmed' | 'rid';
    idtypeen: number;
    value: string;
}

interface Author {
    name: string;
    authtype: string;
    clusterid: string;
}

interface ResultUIDs {
    uids: string[];
}

interface ResultItems {
    [pmid: string]: Eutils.Entry;
}

type Result = ResultUIDs & ResultItems;

declare namespace Eutils {
    interface Response {
        header: {
            type: string;
            version: string;
        };
        result: Result;
        error?: string;
    }
    type Entry = EntryOk | EntryError;
    interface EntryOk {
        __astrocite_kind: 'entry';
        uid: string;
        /** Example: "1975 May" */
        pubdate: ShortDate;
        epubdate: ShortDate;
        /** container-title-short */
        source: string;
        authors: Author[];
        lastauthor: string;
        title: string;
        sorttitle: string;
        volume: string;
        issue: string;
        /** Range (e.g. "525-30") */
        pages: string;
        lang: string[];
        nlmuniqueid: string;
        issn: string;
        essn: string;
        pubtype: string[];
        recordstatus: string;
        /**
         * (1) received      -- date manuscript received for review
         * (2) accepted      -- accepted for publication
         * (3) epublish      -- published electronically by publisher
         * (4) ppublish      -- published in print by publisher
         * (5) revised       -- article revised by publisher/author
         * (6) pmc           -- article first appeared in PubMed Central
         * (7) pmcr          -- article revision in PubMed Central
         * (8) pubmed        -- article citation first appeared in PubMed
         * (9) pubmedr       -- article citation revision in PubMed
         * (10) aheadofprint -- epublish, but will be followed by print
         * (11) premedline   -- date into PreMedline status
         * (12) medline      -- date made a MEDLINE record
         * (255) other
         */
        pubstatus?:
            | '1'
            | '2'
            | '3'
            | '4'
            | '5'
            | '6'
            | '7'
            | '8'
            | '9'
            | '10'
            | '11'
            | '12'
            | '255';
        articleids: ArticleID[];
        history: HistoryEntry[];
        references: Reference[];
        attributes: never[]; // FIXME: not sure what this is
        pmcrefcount: string;
        /** Journal name in sentence case */
        fulljournalname: string;
        elocationid: string;
        doctype: string; // FIXME: what are available?
        srccontriblist: never[]; // FIXME: not sure what this is
        booktitle: string;
        medium: string;
        edition: string;
        publisherlocation: string;
        publishername: string;
        srcdate: LongDate;
        reportnumber: string;
        availablefromurl: string;
        locationlabel: string;
        doccontriblist: never[]; // FIXME: not sure what this is
        docdate: LongDate;
        bookname: string;
        chapter: string;
        sortdate: LongDate; // Used with PMC json
        sortpubdate: LongDate;
        sortfirstauthor: string;
        vernaculartitle: string;
        error: ''; // Does not actually exist here. Needed for discriminated union
    }

    interface EntryError {
        __astrocite_kind: 'error';
        uid: string;
        error: string;
    }
}